requirements.txt
# Azure IoT Hub Device Simulator - Python Dependencies
# This file specifies all required Python packages for the application

# Core GUI Framework
PySide6>=6.5.0

# Azure IoT Hub Integration
azure-iot-device>=2.12.0
azure-identity>=1.13.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
faker>=18.0.0

# System Monitoring and Performance
psutil>=5.9.0

# Data Export and Reporting
openpyxl>=3.1.0
reportlab>=4.0.0

# JSON Schema and Validation
jsonschema>=4.17.0

# Date and Time Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# HTTP Requests and Networking
requests>=2.31.0
urllib3>=2.0.0

# Development and Testing (Optional)
pytest>=7.3.0
pytest-qt>=4.2.0
pytest-asyncio>=0.21.0

# Code Quality Tools (Optional)
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Application Packaging (Optional)
pyinstaller>=5.10.0

